+++ FE/src/components/AccountBook/AccountBookAddForm/CreateForm/create-form.scss
@@ -1,0 +1,16 @@
+.input__name {
+  position: absolute;
+  top: 1.3em;
+  height: 2em;
+  width: 260px;
+  text-decoration: none;
+  margin: 5px;
+}
+.input__description {
+  position: absolute;
+  top: 5em;
+  height: 5em;
+  width: 260px;
+  margin: 5px;
+  resize: none;
+}
