+++ FE/src/components/AccountBook/AccountBookAddForm/JoinForm/join-form.scss
@@ -1,0 +1,9 @@
+.input__code {
+  position: absolute;
+  top: 50%;
+  transform: translateY(-50%);
+  height: 2em;
+  width: 260px;
+  text-decoration: none;
+  margin: 5px;
+}
