+++ FE/src/components/AccountBook/AccountBookAddForm/index.scss
@@ -16,33 +16,6 @@
     color: $spending-color;
     cursor: pointer;
   }
-  .input__name {
-    position: absolute;
-    top: 1.3em;
-    height: 2em;
-    width: 260px;
-    text-decoration: none;
-    margin: 5px;
-  }
-
-  .input__code {
-    position: absolute;
-    top: 50%;
-    transform: translateY(-50%);
-    height: 2em;
-    width: 260px;
-    text-decoration: none;
-    margin: 5px;
-  }
-
-  .input__description {
-    position: absolute;
-    top: 5em;
-    height: 5em;
-    width: 260px;
-    margin: 5px;
-    resize: none;
-  }
 
   @keyframes cardShow {
     from {
